hacky but working tests
