+++ server/fadzmaq/__init__.py
@@ -1,0 +1,4 @@
+
+
+
+from fadzmaq.fadzmaq import create_app
