+++ server/fadzmaq/database/__init__.py
@@ -11,7 +11,10 @@
 # Copyright FadZmaq © 2019      All rights reserved.
 # @author Lachlan Russell       22414249@student.uwa.edu.au
 from sqlalchemy import create_engine
-from database import db_conf, db
+
+# from fadzmaq.database import db_conf, db
+# from database import db_conf, db
+from . import db_conf, db
 
 engine = create_engine(db_conf.DATABASE_URI)
 connection = engine.connect()
