+++ server/fadzmaq/database/build_test.sql
