+++ server/fadzmaq/database/create_server_role.sql
