+++ server/fadzmaq/database/create_test_user.sql
