+++ server/fadzmaq/database/init.sql
