+++ server/run_server_windows.bat
@@ -1,3 +1,3 @@
 @echo on
 start /wait /B pipenv install --dev
-start /wait /B pipenv run python fadzmaq.py
+start /wait /B pipenv run python fadzmaq/fadzmaq.py
